Merged Makefile and module.mk into a single Makefile.
Now the Makefile recursively includes itself, but an ifndef inclusion guard is
used to prevent the "top-level" part of the Makefile from being evaluated more
than once. Added comments to the code that came from module.mk.

Also fixed a bug that resulted in ALL_DEPS having only the last modules'
dependency files in it. This would have caused broken dependencies as well as
incomplete cleaning.