Added distinction between prerequesite libraries (PRELIBS) and regular LIBS.
PRELIBS is used to declare libraries that are prerequesites of the current
target (i.e. these libraries will be put on the target's prerequesite line). As
such, make will attempt to build the libraries listed in PRELIBS and,
accordingly, such libraries must have rules for building them (i.e. they ought
be specified as TARGETs somewhere in the makefile hierarchy).

Regular LIBS are not added as prerequesites, so make will not attempt to build
them. They will simply be added as linker directives (e.g. -l<name>).