Fixed linking so that the linker front-end used depends on the sources.
If C++ sources are present, then it will use the C++ compiler (e.g. g++) as the
front-end for linking. Otherwise, it will use the C compiler.

This can be overriden in the user's modules by specifying LNK := <linker>.

In order to pull off this new capability, it's necessary for the user to
specify the *sources* to be compiled, rather than the object files to be
generated. This is because it is otherwise very difficult (impossible, maybe?)
to figure out what sources were actually used to generate the objects when
pattern rules are in use.