Made pattern rules use the source file extension lists.
This adds a new parameterized function that allows pattern rules to be
generated. This makes it possible to generate the list of pattern rules from the
list of supported sourc file extensions. This way adding support for a new
source file extension is literally as simple as adding it to the appropriate
list of extensions.