Changed PRELIBS to PREREQS to solve library-ordering problem at link time.
There needs to be a way to specify the order that libraries are mentioned on the
linker command line, because the order is significant for resolving
dependencies. But with separate LIBS and PRELIBS, there is no convenient way to
say what order the libraries should be specified on the command line.

Now PRELIBS is just PREREQS: it only controls what targets are prerequesites of
the current target. Prerequesite libraries additionally need to be listed in
LIBS. This allows the exact order of libraries to be specified in LIBS.