Fixed linking problems related to linker selection and shared objects.
Different targets may require different linkers depending on the sources that
make up each target. Previously, the linker was selected for all targets based
upon all sources. So you might end up using the C++ compiler for all targets
even if some of the targets only have C sources.

Also, in order to properly support shared objects, it must be possible to
specify linker flags on a per-target basis (so that .so targets can get the
-shared, -soname, and other related options).

Finally, removed automatic manipulation of the linker search path when library
targets are defined. This gets too sticky with shared objects and then begs the
question as to why other linker flags (like -shared) don't automatically get
specified. It's better to leave this up to the user to specify.