Changed terminology and variable names for consistency and understandability.
All variables that are specific to targets are prefixed with "TGT_" and all
variables that are specific to source files are prefixed with "SRC_". So-called
"global" variables simply go by their standard names (e.g. CFLAGS, LDFLAGS).

This makes it easy to understand what a given variable applies to and also
maintains the normal ability to override the standard variables (e.g. from the
command line).

Also, ditched the "module" moniker in favor of "submakefile", which seems more
straightforward (albeit a bit longer).