Added per-target "post-clean" variables; made SRC_INCDIR relative to *.mk dir.
The old POSTCLEAN variable is now obsolete and should not be used in any
submakefile. Instead, each target can specify its own set of postclean actions.
This allows multiple submakefiles to contain a postclean directive without them
interfering with each other.

Also, this makes the SRC_INCDIRS variable specification relative to the
directory that the submakefile itself is in (rather than the top-level
directory). This makes it possible for superprojects to include submakefiles
from subprojects. Before, doing this would mean that the top-level directory
changes (to the superproject's directory) and this causes all SRC_INCDIR
variables in subprojects to now refer to incorrect paths.