Strip echoed command line in command scripts
Often, command lines echoed to stdout during make's execution contain
extraneous whitespace. These are extra spaces that get left behind when a
variable in one of the commands is empty, resulting in a couple of spaces
surrounding the empty string. This change strips all command lines that are
echoed to stdout so that any extraneous spaces are removed.