Rename main.mk -> MANUAL, and mention MANUAL in README
Putting the boilermake documentation in main.mk might be a bit confusing or
even misleading. So, main.mk has been renamed MANUAL. Also, the previous
version of the README file neglected to mention the documentation for all of
the boilermake special variables. The README file now refers the reader to
MANUAL for variable documentation.