Redesign the test-app submakefiles to demonstrate "subprojects"
Since around July 2009, boilermake has supported a submakefile structure such
that the make command could be executed in designated subdirectories in order
to build just subprojects of the overall project (e.g. you could build or
re-build just a single library without having to rebuild all of the other
libraries and/or executables that are defined as targets elsewhere in the
submakefile hierarchy).

However, a method for doing this was never documented and it isn't exactly
obvious either. This change redesigns the test-app so that it serves as a
demonstration of how submakefiles can be arranged in order to make building
subprojects possible.

Now, the "animals" project (which generates the libanimals.a static library) is
effectively a "subproject" of the main "talk" project. It's possible to run
"make" or "make clean" from within the "animals" directory and *only* the
libanimals.a library gets rebuilt.