Fix version requirement error message if not using GNU Make (or ancient version)
Other make systems (or possibly even ancient versions of GNU Make) don't define
the MAKE_VERSION variable. If that's the case, we need to print an alternative
informational message, otherwise we print just "This is GNU Make version" and
nothing else which is odd, and would actually be wrong in cases where the
installed make system is *not* GNU make.