Stop using target-specific variables for TGT_* variables
One property of target-specific variables is that they can be inherited by
prerequisites. This could cause, for example, TGT_LDFLAGS for target A to
theoretically be inherited by another target, B. But B should only use
TGT_LDFLAGS that were explicitly set for it. At present, this isn't a problem
for boilermake because all of the TGT_* variables are always explicitly set
(at least to the empty string if nothing else) for each target, which overrides
the inheritance mechanism. However, it's fragile and could potentially lead
to problems later. Also, since some of the variables, such as TGT_PREREQS, can't
use target-specific variables anyway (because they are referenced outside of the
target's command script), there was a mix of some target-specific variables and
some regular variables with the target's name prefixed to their names. It will
be more consistent if all variables are regular variables with the target's
name prefixed (i.e. ${TGT}_FOO).